+++ src/EntityGraphQL/Schema/ISchemaType.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Linq.Expressions;
 
 namespace EntityGraphQL.Schema
 {
@@ -19,9 +18,13 @@ namespace EntityGraphQL.Schema
         void AddField(Field field);
         void RemoveField(string name);
         /// <summary>
-        /// Add all fields and properties from the dotnet type to the schema type
+        /// Add all public Properties and Fields from the DotNet type to the schema type.
         /// </summary>
+        /// <param name="schema">The schema to add new Types to</param>
+        /// <param name="autoCreateNewComplexTypes">Default false. If true creates new schema types for any complex dotnet types found.false Also adding all it's fields</param>
+        /// <param name="autoCreateEnumTypes">Default true. If true creates new Enum types in the schema for any enums found as field types</param>
+        /// <typeparam name="TContextType"></typeparam>
         /// <returns></returns>
-        ISchemaType AddAllFields();
+        ISchemaType AddAllFields<TContextType>(MappedSchemaProvider<TContextType> schema, bool autoCreateNewComplexTypes = false, bool autoCreateEnumTypes = true);
     }
 }