+++ src/EntityGraphQL/Schema/MappedSchemaProvider.cs
@@ -36,13 +36,13 @@ namespace EntityGraphQL.Schema
             queryContextName = queryContext.Name;
             types.Add(queryContext.Name, queryContext);
 
-            AddType<Models.InputValue>("__InputValue", "Arguments provided to Fields or Directives and the input fields of an InputObject are represented as Input Values which describe their type and optionally a default value.").AddAllFields();
-            AddType<Models.Directives>("__Directive", "Information about directives").AddAllFields();
-            AddType<Models.EnumValue>("__EnumValue", "Information about enums").AddAllFields();
-            AddType<Models.Field>("__Field", "Information about fields").AddAllFields();
-            AddType<Models.Schema>("__Schema", "A GraphQL Schema defines the capabilities of a GraphQL server. It exposes all available types and directives on the server, as well as the entry points for query, mutation, and subscription operations.").AddAllFields();
-            AddType<Models.SubscriptionType>("Information about subscriptions").AddAllFields();
-            AddType<Models.TypeElement>("__Type", "Information about types").AddAllFields();
+            AddType<Models.TypeElement>("__Type", "Information about types").AddAllFields(this);
+            AddType<Models.EnumValue>("__EnumValue", "Information about enums").AddAllFields(this);
+            AddType<Models.InputValue>("__InputValue", "Arguments provided to Fields or Directives and the input fields of an InputObject are represented as Input Values which describe their type and optionally a default value.").AddAllFields(this);
+            AddType<Models.Directives>("__Directive", "Information about directives").AddAllFields(this);
+            AddType<Models.Field>("__Field", "Information about fields").AddAllFields(this);
+            AddType<Models.SubscriptionType>("Information about subscriptions").AddAllFields(this);
+            AddType<Models.Schema>("__Schema", "A GraphQL Schema defines the capabilities of a GraphQL server. It exposes all available types and directives on the server, as well as the entry points for query, mutation, and subscription operations.").AddAllFields(this);
 
             Type<Models.TypeElement>("__Type").ReplaceField("enumValues", new { includeDeprecated = false },
                 (t, p) => t.EnumValues.Where(f => p.includeDeprecated ? f.IsDeprecated || !f.IsDeprecated : !f.IsDeprecated).ToList(), "Enum values available on type");
@@ -222,7 +222,8 @@ namespace EntityGraphQL.Schema
         /// <returns></returns>
         public SchemaType<TType> Type<TType>()
         {
-            return (SchemaType<TType>)types[typeof(TType).Name];
+            // look up by the actual type not the name
+            return (SchemaType<TType>) types.Values.Where(t => t.ContextType == typeof(TType)).First();
         }
         public SchemaType<TType> Type<TType>(string typeName)
         {
@@ -539,7 +540,7 @@ namespace EntityGraphQL.Schema
         {
             var schemaType = new SchemaType<object>(type, name, description, false, true);
             types.Add(name, schemaType);
-            return schemaType.AddAllFields();
+            return schemaType.AddAllFields(this);
         }
     }
 }